Recreated app for .NET Core 3.1
